Correctly load the apm agent executable. Make platform-specific locations for the executable (the relative path is different from windows <--> linux, as is the executable name
